feat: Code cleanup and type safety improvements
This commit introduces several changes to improve code organization, reduce redundancy, and enhance type safety across the application.

Key changes include:

1.  **Directory Structure & Naming:**
    *   Initially renamed `app/profile/ItemCard.tsx` for clarity, which I subsequently removed after consolidation.

2.  **Code Reusability & Reduction:**
    *   I consolidated two `ItemCard` components into a single, more versatile `components/ItemCard.tsx`. This involved merging props and functionality for displaying item details and user-specific actions (edit, delete, etc.).
    *   I removed the redundant `app/profile/ProfileItemCard.tsx`.
    *   I updated profile pages (`app/profile/page.tsx`, `app/profile/[id]/page.tsx`) to use the unified `ItemCard`.
    *   I refactored the homepage (`app/page.tsx`) to use the `ItemCard` for displaying featured items, promoting consistency and reducing code duplication.
    *   I removed unused imports from `app/page.tsx`.

3.  **Type Definitions & Safety:**
    *   I corrected a typo in the `Message` interface in `lib/types.ts` (reciever_id -> receiver_id).
    *   I expanded the `Item` interface in `lib/types.ts` with optional fields (`sender_id`, `seller`) to better align with data access patterns in `hooks/useItem.ts`.
    *   I strengthened typing in `hooks/useItem.ts`:
        *   I typed the `seller` state as `User | null` instead of `any`.
        *   I improved the logic for determining `sellerId` to be more explicit and type-aware.
        *   I ensured fetched seller data is cast to the `User` type.
    *   I confirmed that `tsconfig.json` already enforces `"strict": true`, which is beneficial for overall type safety.

These changes contribute to a cleaner, more maintainable, and more robust codebase.
